ref(upload): Move ExactStream to utils and make it generic
Move ExactStream from the upload service module to relay-server utils,
making it generic over the inner stream type. The Stream impl requires
S: Unpin, which is naturally satisfied by boxed streams (Pin<Box<…>>).
